11_TP_Deployment_Ansible/Ansible/files/requirements.txt
opcua
paho-mqtt
cryptography
PyYAML